Refactor: 폴더 구조 변경
- components에는 공통으로 혹은 재사용이 필요한 컴포넌트만 관리
- 단순히 코드 분리를 위한 컴포넌트는 templates의 각 섹션에서 관리
